Pipeline working on offline data tested

+++ src/utilities/Utils.h
@@ -24,7 +24,7 @@ QDateTime MJD2QDateTime(const double inMJD);
  * @brief UnixTime2MJD
  * Converts a unix time to modified julian date
  */
-quint32 UnixTime2MJD(const double inUnixTime);
+double UnixTime2MJD(const double inUnixTime);
 
 /**
  * @brief MJD2UnixTime
